fix(chat): render the Add-Credits CTA for thread-parented pause notices (#1846)
A budget-pause notice is journaled with the thread's parent set when
the exhausted turn was answering a reply, and buildTimelineItems
routes any message with a parentId out of the main channel timeline
and into ThreadPanel. The redeem callback was wired only into
MessageTimeline/MessageRow, so ThreadPanel rendered every system
message (including this one) as plain text with no CTA — the only
place such a notice was visible offered no way to redeem it.

Extracted MessageRow's SystemPill notice-card JSX into a shared
BudgetPauseNoticeCard component and wired it into ThreadPanel's own
system-message rendering, threading onRedeemBudgetPause/
redeemingBudgetPauseAgent/latestBudgetPauseMessageIdByAgent down from
ChatView the same way MessageTimeline already receives them.
