+++ frontend/src/views/chat/ThreadPanel.tsx
@@ -7,6 +7,7 @@ import type { MessageIntent } from "@/api/tasks";
 import type { AttachmentDto, CognitionState } from "@/api/types";
 import type { ChatMessage } from "@/lib/chat";
 import type { TeamMember } from "@/lib/team";
+import { BudgetPauseNoticeCard } from "./BudgetPauseNoticeCard";
 import { EchoPlaceholder, echoMarkerFor } from "./EchoPlaceholder";
 import { MessageAttachments } from "./MessageAttachments";
 import { MessageComposer } from "./MessageComposer";
@@ -94,6 +95,18 @@ interface Props {
    * it had been removed (CodeRabbit and codex both caught it on PR #1740).
    */
   cognition?: CognitionState | null;
+  /**
+   * The Add-Credits CTA (issue #1846 review, Codex #3870168372). A
+   * budget-pause notice is journaled with the thread's `parent` set when the
+   * exhausted turn was answering a reply, and `buildTimelineItems` routes any
+   * message with a `parentId` out of the main channel timeline and into this
+   * thread — so a thread-parented notice's CTA has to render HERE, not just
+   * in `MessageTimeline`, or it is unreachable no matter how the operator
+   * scrolls the main channel.
+   */
+  onRedeemBudgetPause?: (agentId: string, noticeMessageId: string) => void;
+  redeemingBudgetPauseAgent?: string | null;
+  latestBudgetPauseMessageIdByAgent?: Map<string, string>;
 }
 
 /**
@@ -118,6 +131,9 @@ export function ThreadPanel({
   typingNames = [],
   onTyping,
   cognition,
+  onRedeemBudgetPause,
+  redeemingBudgetPauseAgent,
+  latestBudgetPauseMessageIdByAgent,
 }: Props) {
   return (
     <aside className="flex w-96 shrink-0 flex-col border-l bg-background">
@@ -139,6 +155,9 @@ export function ThreadPanel({
           youAvatar={youAvatar}
           resolveAttachmentUrl={resolveAttachmentUrl}
           cognition={cognition}
+          onRedeemBudgetPause={onRedeemBudgetPause}
+          redeemingBudgetPauseAgent={redeemingBudgetPauseAgent}
+          latestBudgetPauseMessageIdByAgent={latestBudgetPauseMessageIdByAgent}
         />
         <div className="flex items-center gap-2 px-4 py-2">
           <span className="text-xs font-medium text-muted-foreground">
@@ -155,6 +174,9 @@ export function ThreadPanel({
             youAvatar={youAvatar}
             resolveAttachmentUrl={resolveAttachmentUrl}
             cognition={cognition}
+            onRedeemBudgetPause={onRedeemBudgetPause}
+            redeemingBudgetPauseAgent={redeemingBudgetPauseAgent}
+            latestBudgetPauseMessageIdByAgent={latestBudgetPauseMessageIdByAgent}
           />
         ))}
       </div>
@@ -180,6 +202,9 @@ function Line({
   youAvatar,
   resolveAttachmentUrl,
   cognition,
+  onRedeemBudgetPause,
+  redeemingBudgetPauseAgent,
+  latestBudgetPauseMessageIdByAgent,
 }: {
   channel: Channel;
   members: TeamMember[];
@@ -187,6 +212,9 @@ function Line({
   youAvatar?: string;
   resolveAttachmentUrl?: (nodeId: string) => Promise<string>;
   cognition?: CognitionState | null;
+  onRedeemBudgetPause?: (agentId: string, noticeMessageId: string) => void;
+  redeemingBudgetPauseAgent?: string | null;
+  latestBudgetPauseMessageIdByAgent?: Map<string, string>;
 }) {
   // Four arguments, not three: `youAvatar` is the last parameter, and omitting
   // it left your own line with no avatar to seed from but the name "You" —
@@ -198,8 +226,26 @@ function Line({
   const placeholder = echoMarkerFor(message, sender, cognition);
 
   if (sender.kind === "system") {
+    // Issue #1846 review (Codex #3870168372): a budget-pause notice gets the
+    // SAME highlighted card + CTA `MessageRow` renders in the main timeline
+    // — see `BudgetPauseNoticeCard`'s own doc for why this thread panel is
+    // sometimes the ONLY place such a notice is visible at all. Falls back
+    // to the plain system line for every other system message.
+    // Called as a plain function, not as JSX: `BudgetPauseNoticeCard` is a
+    // pure "maybe-render" component (no hooks, no state) that returns `null`
+    // for a non-notice message — JSX-invoking it (`<BudgetPauseNoticeCard
+    // .../>`) would always produce a truthy element descriptor regardless of
+    // what it renders internally, defeating the `??` fallback below.
+    const budgetPauseCard = BudgetPauseNoticeCard({
+      message,
+      onRedeemBudgetPause,
+      redeemingBudgetPauseAgent,
+      latestBudgetPauseMessageIdByAgent,
+    });
     return (
-      <p className="px-4 py-1 text-center text-xs text-muted-foreground">{message.text}</p>
+      budgetPauseCard ?? (
+        <p className="px-4 py-1 text-center text-xs text-muted-foreground">{message.text}</p>
+      )
     );
   }
